google_adk_kit/app/requirements.txt
# SHL Assessment Recommendation System - Requirements

# FastAPI and Web Server
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9

# Google ADK (Agent Development Kit)
google-adk==0.1.0

# Vector Search and Embeddings
faiss-cpu==1.8.0
sentence-transformers==3.1.1

# Data Processing
numpy==1.26.4
pandas==2.2.2

# Environment and Configuration
python-dotenv==1.0.1

# HTTP Client (for testing)
httpx==0.27.2
